Safari-611.1.10/LayoutTests/media/modern-media-controls/tracks-support/tracks-support-no-tracks-expected.txt

Testing the TracksSupport behavior with no tracks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.tracksButton.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
